# Optimize parser by removing repeated hash merges