@bkeepers 2024-12-12
@andrewts129 awesome, thanks for finding this and working on a fix!

I pushed some benchmarks in 1877fa0 just to compare. Here's the results of loading a 1000 line `.env` file:

**main**

```
parse, overwrite:false
                         43.265 (±11.6%) i/s   (23.11 ms/i) -    196.000 in   5.056455s
parse, overwrite:true
                         26.042 (± 0.0%) i/s   (38.40 ms/i) -    132.000 in   5.069958s
```

**this branch**

```
parse, overwrite:false
                        552.652 (± 0.9%) i/s    (1.81 ms/i) -      2.805k in   5.075897s
parse, overwrite:true
                        569.266 (± 0.7%) i/s    (1.76 ms/i) -      2.850k in   5.006666s
```

13-21x faster. Nice work!