@idrozd 2025-05-26
Jesus

Thanks @andrewts129 , I thought I was going nuts

Spent hours and hours figuring this out, ending up with monkeypatch.

I could swear ENV#each / #to_h was always fast until it wasn't sometime last year.